refactor(types): move API types into types.rs
