+++ src/lib.rs
@@ -10,4 +10,4 @@ mod error;
 mod helpers;
 
 pub mod api;
-pub mod responses;
+pub mod types;
